% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANGUAGE.R
\docType{data}
\name{LANGUAGE}
\alias{LANGUAGE}
\title{LANGUAGE}
\format{
A data frame with 11 rows and 11 columns:
\describe{
  \item{E}{English}
  \item{N}{Norwegian}
  \item{Da}{Danish}
  \item{Du}{Dutch}
  \item{G}{German}
  \item{Fr}{French}
  \item{Sp}{Spanish}
  \item{I}{Itialian}
  \item{P}{Polish}
  \item{H}{Hungarian}
  \item{Fi}{Finnish}
  ...
}
}
\source{
https://www.amazon.com/Multivariate-Statistical-Analysis-Classics-Statistics...
}
\usage{
LANGUAGE
}
\description{
A data set of concordant first letters for numbers in 11 languages
}
\keyword{datasets}
