% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maleturt.R
\docType{data}
\name{maleturt}
\alias{maleturt}
\title{maleturt}
\format{
A data frame with 24 rows and 3 columns:
\describe{
  \item{LENGTH}{Legnth of carapace}
  \item{WIDTH}{Width of carapace}
  \item{HEIGHT}{Height of carapace}
  ...
}
}
\source{
<https://www.amazon.com/Statistics-Engineering-Sciences-William...
}
\usage{
maleturt
}
\description{
A data set of measurements of 24 male turtles' carapaces
}
\keyword{datasets}
