% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro.R
\docType{data}
\name{micro}
\alias{micro}
\title{MICROWAVE}
\format{
A data frame with 42 rows and 2 columns:
\describe{
  \item{Closed}{Radiation measures with door closed}
  \item{Open}{Radiation measures with door open}
  ...
}
}
\source{
https://www.amazon.com/Multivariate-Statistical-Analysis-Classics-Statistics...
}
\usage{
micro
}
\description{
A data set of the amount of radiation that 42 microwave ovens emit with the door closed and open
}
\keyword{datasets}
