% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTSQint.R
\name{myTSQint}
\alias{myTSQint}
\title{myTSQint}
\usage{
myTSQint(df, alpha = 0.05)
}
\arguments{
\item{df}{bivariate data frame}

\item{alpha}{alpha level for ci's}
}
\value{
confidence intervals and plot of the contour ellipse
}
\description{
A function which calculates the simultaneous T-squared intervals for component means
}
\examples{
\dontrun{myTSQint(df = micr, alpha = 0.05)}
}
