% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLYMPIC.R
\docType{data}
\name{OLYMPIC}
\alias{OLYMPIC}
\title{OLYMPIC}
\format{
A data frame with 10 rows and 10 columns:
\describe{
  \item{V1}{Variable 1}
  \item{V2}{Variable 2}
  \item{V3}{Variable 3}
  \item{V4}{Variable 4}
  \item{V5}{Variable 5}
  \item{V6}{Variable 6}
  \item{V7}{Variable 7}
  \item{V8}{Variable 8}
  \item{V9}{Variable 9}
  \item{V10}{Variable 10}
  ...
}
}
\source{
https://www.amazon.com/Multivariate-Statistical-Analysis-Classics-Statistics...
}
\usage{
OLYMPIC
}
\description{
A data set of Olympic decathalon results, this is a correlation matrix
}
\keyword{datasets}
