% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfactors.R
\name{shinyfactors}
\alias{shinyfactors}
\title{shinyfactors}
\usage{
shinyfactors()
}
\value{
a table of principal component solutions for the given number of factors
}
\description{
An interactive shiny app which allows the choice of the number of factors for factor analysis
}
\examples{
\dontrun{shinypointest()}
}
